+++ py2app/recipes/__init__.py
@@ -12,3 +12,4 @@ from . import matplotlib
 from . import virtualenv
 from . import pyside
 from . import wx
+from . import pyzmq